+++ plugins/org.ietr.preesm.mapper/src/org/ietr/preesm/mapper/graphtransfo/SDF2DAGTransformation.java
@@ -1,0 +1,74 @@
+/**
+ * 
+ */
+package org.ietr.preesm.mapper.graphtransfo;
+
+import java.util.LinkedHashMap;
+import java.util.Map;
+import org.eclipse.core.runtime.IProgressMonitor;
+import org.ietr.dftools.algorithm.model.sdf.SDFGraph;
+import org.ietr.dftools.algorithm.model.visitors.SDF4JException;
+import org.ietr.dftools.architecture.slam.Design;
+import org.ietr.dftools.workflow.WorkflowException;
+import org.ietr.dftools.workflow.elements.Workflow;
+import org.ietr.dftools.workflow.implement.AbstractTaskImplementation;
+import org.ietr.dftools.workflow.implement.AbstractWorkflowNodeImplementation;
+import org.ietr.preesm.core.scenario.PreesmScenario;
+import org.ietr.preesm.mapper.model.MapperDAG;
+
+/**
+ * Class used to transform a Single-Rate SDF graph to a Directed Acyclic Graph
+ * 
+ * @author farresti
+ *
+ */
+public class SDF2DAGTransformation extends AbstractTaskImplementation {
+
+  /*
+   * (non-Javadoc)
+   * 
+   * @see org.ietr.dftools.workflow.implement.AbstractTaskImplementation#execute(java.util.Map, java.util.Map, org.eclipse.core.runtime.IProgressMonitor,
+   * java.lang.String, org.ietr.dftools.workflow.elements.Workflow)
+   */
+  @Override
+  public Map<String, Object> execute(Map<String, Object> inputs, Map<String, String> parameters, IProgressMonitor monitor, String nodeName, Workflow workflow)
+      throws WorkflowException {
+    final SDFGraph algorithm = (SDFGraph) inputs.get(AbstractWorkflowNodeImplementation.KEY_SDF_GRAPH);
+    final PreesmScenario scenario = (PreesmScenario) inputs.get(AbstractWorkflowNodeImplementation.KEY_SCENARIO);
+    final Design architecture = (Design) inputs.get(AbstractWorkflowNodeImplementation.KEY_ARCHITECTURE);
+
+    final MapperDAG dag;
+    try {
+      dag = SdfToDagConverter.convert(algorithm, architecture, scenario);
+    } catch (SDF4JException e) {
+      throw (new WorkflowException(e.getMessage()));
+    }
+    // TODO Auto-generated method stub
+    final Map<String, Object> outputs = new LinkedHashMap<>();
+    outputs.put(AbstractWorkflowNodeImplementation.KEY_SDF_DAG, dag);
+    return outputs;
+  }
+
+  /*
+   * (non-Javadoc)
+   * 
+   * @see org.ietr.dftools.workflow.implement.AbstractTaskImplementation#getDefaultParameters()
+   */
+  @Override
+  public Map<String, String> getDefaultParameters() {
+    // TODO Auto-generated method stub
+    return null;
+  }
+
+  /*
+   * (non-Javadoc)
+   * 
+   * @see org.ietr.dftools.workflow.implement.AbstractWorkflowNodeImplementation#monitorMessage()
+   */
+  @Override
+  public String monitorMessage() {
+    // TODO Auto-generated method stub
+    return null;
+  }
+
+}